avoid use of nodes for rendering
This could otherwise lead to conflicts if some zooms/rotations use nodes
and subsequent zooms/rotations or later renders/previews do not use
nodes.

Depending on whether nodes are used for anything else in the future,
this may need to be changed so that all zooms/rotations do use nodes,
but for now there is no need for nodes.