correct orthographic scale for all zooms
Especially for zooms 1 to 3, the orthographic scale was quite a bit off.
For zooms 4 and 5, it is more precise now as well.

The general approach follows code contained in BAT4Max 5, using a 16×16
reference cell, which renders with specific widths in pixels.

The results were verified using a pixel grid as UV-map, which resulted
in a pixel-perfect display in game.