fix denoising artifacts due to tiled rendering
The fix is to render the full image (instead of 256×256 tiles) to avoid
artifacts introduced by denoising filter at the border of 256×256 tiles
when these are rendered one-by-one.

This now renders the full image, applies a denoising filter to the full
image if any, and cuts the image into 256×256 pieces as a final
post-processing step.